Everything is working but validation dont work

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/AdminController.java
@@ -3,10 +3,13 @@ package ru.kata.spring.boot_security.demo.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import ru.kata.spring.boot_security.demo.model.User;
 import ru.kata.spring.boot_security.demo.service.RoleService;
 import ru.kata.spring.boot_security.demo.service.UserService;
+
+import javax.validation.Valid;
 
 @Controller
 public class AdminController {
@@ -38,7 +41,12 @@ public class AdminController {
     }
 
     @PostMapping("/addNewUser")
-    public String saveUser(@ModelAttribute("user") User user, Model model) {
+    public String saveUser(@ModelAttribute("user") @Valid User user,
+                           BindingResult bindingResult, Model model) {
+
+        if (bindingResult.hasErrors()) {
+            return "new";
+        }
 
         model.addAttribute("users", new User());
         model.addAttribute("roles", roleService.getAllRoles());
@@ -73,7 +81,13 @@ public class AdminController {
     }
 
     @PatchMapping("/editUser/{id}")
-    public String edit(@ModelAttribute("user") User user, @PathVariable("id") Long id) {
+    public String edit(@ModelAttribute("user") @Valid User user,
+                       BindingResult bindingResult,
+                       @PathVariable("id") Long id) {
+
+        if (bindingResult.hasErrors()) {
+            return "edit";
+        }
 
         userService.updateUser(id, user);
 
@@ -83,9 +97,9 @@ public class AdminController {
    @DeleteMapping("/{id}")
     public String deleteUser(@PathVariable("id") Long id) {
 
-//        if (userService.getUserById(id) == null) {
-//            return "error-page";
-//        }
+        if (userService.getUserById(id) == null) {
+            return "error-page";
+        }
         userService.deleteUser(id);
 
         return "redirect:/admin";
